// src/models/SocietyMemberTenantsModel.tsx
export interface SocietyMemberTenantsModel {
    SocietyMemberTenantId: string;
    SocietyId: string;
    FolioNo: number;
    Tenant: string;
    StartDate: Date;
    EndDate : Date;
    Address: string;
    City: string;
    StateId: string;
    Pin: number;
    CountryCode: Date;
    PhoneNo: string;
    OfficePhoneNo: string;
    MobileNo: string;
    EmailId: string;
    IsActive: boolean;
    ApprovedById: string;
    ApprovedByName: string;
    ApprovedOn: Date;
    
    
  } 